Tweaks dependabot to use `-` as separator
Because we need to push and pull images from a registry, branches with
`/` in the name are really hard to support and that brakes the
namespaces and similar. Thankfully, dependabot can be configured to use
a different separator character.

Docs: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#pull-request-branch-nameseparator