Add PAT permissions to write to packages
According to the docs (see below), we can specify we can use the
`permission` keyword to modify the `GITHUB_TOKEN` granted privileges.
This should help dependabot being able to push to the package registry
because we need this for caching purposes (we're caching things in some
stages to speed up the pipeline).

Docs: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#permissions