dns_ips: Deprecate iprangethirdparty w tags in iprange
IPRangeThirdParty is an old custom model that no longer makes sense as
the same concept can be achieved by tagging `IPRange` properly.
This change adds `tags` field to the `IPRange` model and tries, in the
same migration, to tag existing ranges accordingly.

It is adivsed in other installations to create a custom migration to
move over data from IPRangeThirdParty to IPRange to avoid losing data.
Even without tags, this migration will iterate *all* `IPRangeThirdParty`
to fetch them in the `IPRange` module to tag them correctly.
Other installations can (or should!?) leverage this new field to create
a separate migration to add tags to the IPRange. Since this
implementation can vary and have different impact on people, this change
should be done by each implementor separately.