dns_ips: Deprecate DNSNameservers
This old model was intended to whitelist some nameservers but they ended
up never being used.
In addition, while reviewing the model, I noticed there was a duplicate
field using the same structure, a many-to-many type of field to the same
model.

To help those with Surface installed remove this foreign key, this change
includes a migration to move the data from the existing fields to a new
one, TextField, comma separated. This helps keeping the data.