feature: add Finding to Surface
`Finding` is a key model in Surface to allow others to bootstrap their
own Vulnerability Management program.
With this simple implementation, users can create classes in other apps
that inherit this one, centralizing vulnerabilities and risks from all
other apps of their own implementation.
Examples to be posted in the Wiki.